Add white list feature
This feature works by providing a list of comma separated URLs that the
tool will not check status of. It works with exact matches and also if
you just provide the start of the URL.

E.g. "google.com" will filter out both "google.com" as well as
"google.com/sub"